Allow creation of free plans
In order to create free plans we have to pass in $0 for the plan’s `UnitAmountInCents`. Unfortunately `UnitAmountInCents` is also used in a bunch of other structs like adjustments, coupons, and subscriptions and I am too worried to change its behavior without fully understanding what the consequences of such change are, so instead I propose to use a new struct that is just used for plans and just sets 0 amount for dollars. It isn’t extensible for the multi-currency world, but it will work for us since we only create plans in USD.

We only modify plans in manual tools, like `tool create-new-leasable-plugin`.